Use one extension for suggestions and action button
